requirements.txt
PyQt5
TikTokAPI
googleapiclient.discovery
google.oauth2.credentials
instagram_api